+++ src/components/App.js
@@ -1,0 +1,13 @@
+import BooksList from '../containers/BooksList';
+import BooksForm from '../containers/BooksForm';
+import CategoryFilter from './CategoryFilter';
+
+const App = () => (
+  <div>
+    <CategoryFilter />
+    <BooksList />
+    <BooksForm />
+  </div>
+);
+
+export default App;
