+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import './index.css';
-import App from './containers/App';
+import App from './components/App';
 import rootReducer from './redux/reducers';
 
 const store = createStore(rootReducer);
